requirement.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
mongoengine==0.29.0
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pymongo==4.8.0
sqlparse==0.5.1
tzdata==2024.1
